+++ line-config-ui/package.js
@@ -1,15 +1,16 @@
 Package.describe({
   name: 'storyteller:line-config-ui',
   summary: 'Blaze configuration templates for LINE OAuth.',
-  version: '1.1.2',
+  version: '1.2.0',
   git: 'https://github.com/StorytellerCZ/meteor-accounts-line'
 });
 
 Package.onUse(api => {
-  api.versionsFrom('1.12');
-  api.use('ecmascript', 'client');
-  api.use('templating@1.4.1', 'client');
+  api.versionsFrom('2.3');
+  api.use('zodern:types', 'server')
+  api.use(['ecmascript', 'typescript'], 'client');
+  api.use('templating', 'client');
 
   api.addFiles('line_login_button.css', 'client');
-  api.addFiles(['line_configure.html', 'line_configure.js'], 'client');
+  api.addFiles(['line_configure.html', 'line_configure.ts'], 'client');
 });