+++ line-oauth/package.js
@@ -11,15 +11,16 @@ Npm.depends({
 
 Package.onUse(api => {
   api.versionsFrom(['2.9']);
-  api.use('ecmascript', ['client', 'server']);
+  api.use('zodern:types', 'server')
+  api.use(['ecmascript', 'typescript'], ['client', 'server']);
   api.use('oauth2', ['client', 'server']);
   api.use('oauth', ['client', 'server']);
   api.use(['fetch', 'url'], 'server');
   api.use('random', 'client');
   api.use('service-configuration', ['client', 'server']);
 
-  api.addFiles('line_client.js', 'client');
-  api.addFiles('line_server.js', 'server');
+  api.addFiles('line_client.ts', 'client');
+  api.addFiles('line_server.ts', 'server');
 
   api.export('Line');
 });