+++ line-oauth/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 1.3.0 - 2022-12-13
 * Compatibility with Meteor 2.9
+* Typescript conversion
 
 ## 1.2.3 - 2022-10-19
 * Added missing imports into the code